+++ backend/app/__init__.py
@@ -1,4 +1,4 @@
 """Mosaico Backend Application"""
 
-__version__ = "0.2.3"
+__version__ = "0.2.4"
 __app_name__ = "mosaico"